+++ service/frontend/workflowHandler.go
@@ -55,7 +55,7 @@ import (
 
 const (
 	getDomainReplicationMessageBatchSize = 100
-	defaultLastMessageID                 = -1
+	defaultLastMessageID                 = int64(-1)
 )
 
 const (
